Reuse common filterNonPermittedItem
As part of a previous refactor this was moved so
this change reuses the moved version
