+++ src/apps/filters.js
@@ -1,5 +1,3 @@
-const { get, intersection } = require('lodash')
-
 function filterDisabledOption ({ currentValue = null, createdOn } = {}) {
   const createdOnTime = Date.parse(createdOn) || Date.now()
 
@@ -15,17 +13,6 @@ function filterDisabledOption ({ currentValue = null, createdOn } = {}) {
   }
 }
 
-function filterNonPermittedItem (userPermissions) {
-  return function (item) {
-    const hasPermissions = get(item, 'permissions.length', 0)
-    if (!hasPermissions) { return true }
-
-    const isPermitted = intersection(item.permissions, userPermissions).length > 0
-    return isPermitted
-  }
-}
-
 module.exports = {
   filterDisabledOption,
-  filterNonPermittedItem,
 }
